+++ src/java/beans/StaffBean.java
@@ -101,10 +101,10 @@ public class StaffBean {
             GroupRightBean grb = new GroupRightBean();
 
             //String sql1 = "SELECT count(*) as n FROM staff WHERE first_name='" + aStaff.getFirst_name() + "'";
-            if (aStaff.getStaff_id() == 0 && grb.IsUserGroupsFunctionAccessAllowed(aCurrentUserDetail, aCurrentGroupRights, "8", "Add") == 0) {
+            if (aStaff.getStaff_id() == 0 && grb.IsUserGroupsFunctionAccessAllowed(aCurrentUserDetail, aCurrentGroupRights, "88", "Add") == 0) {
                 msg = "Not Allowed to Access this Function";
                 FacesContext.getCurrentInstance().addMessage("Save", new FacesMessage(ub.translateWordsInText(BaseName, msg)));
-            } else if (aStaff.getStaff_id() > 0 && grb.IsUserGroupsFunctionAccessAllowed(aCurrentUserDetail, aCurrentGroupRights, "8", "Edit") == 0) {
+            } else if (aStaff.getStaff_id() > 0 && grb.IsUserGroupsFunctionAccessAllowed(aCurrentUserDetail, aCurrentGroupRights, "88", "Edit") == 0) {
                 msg = "Not Allowed to Access this Function";
                 FacesContext.getCurrentInstance().addMessage("Save", new FacesMessage(ub.translateWordsInText(BaseName, msg)));
             } else if (aStaff.getFirst_name().length() == 0) {
@@ -197,28 +197,40 @@ public class StaffBean {
         }
         return IsUpdated;
     }
-
+    
     public int deleteStaff(Staff aStaff) {
-        int IsDeleted = 0;
         UtilityBean ub = new UtilityBean();
         String BaseName = "language_en";
-        try {
-            BaseName = menuItemBean.getMenuItemObj().getLANG_BASE_NAME_SYS();
-        } catch (Exception e) {
-        }
         String msg = "";
-        String sql = "DELETE FROM staff WHERE staff_id=?";
-        try (
-                Connection conn = DBConnection.getMySQLConnection();
-                PreparedStatement ps = conn.prepareStatement(sql);) {
-            ps.setInt(1, aStaff.getStaff_id());
-            ps.executeUpdate();
-            IsDeleted = 1;
-            this.clearStaff(aStaff);
-            msg = "Staff Deleted Successfully";
+        int IsDeleted = 0;
+        long N = 0;
+        try {
+            String sqlFind = "SELECT COUNT(*) AS n FROM timesheet WHERE staff_id=" + aStaff.getStaff_id();
+            N = N + new UtilityBean().getN(sqlFind);
         } catch (Exception e) {
             LOGGER.log(Level.ERROR, e);
-            msg = "Staff NOT Deleted";
+        }
+        UserDetail aCurrentUserDetail = new GeneralUserSetting().getCurrentUser();
+        List<GroupRight> aCurrentGroupRights = new GeneralUserSetting().getCurrentGroupRights();
+        GroupRightBean grb = new GroupRightBean();
+
+        if (aStaff.getStaff_id() > 0 && grb.IsUserGroupsFunctionAccessAllowed(aCurrentUserDetail, aCurrentGroupRights, "88", "Delete") == 0) {
+            msg = "Not Allowed to Access this Function";
+        } else if (N > 0) {
+            msg = "Staff has been used and cannot be Deleted";
+        } else {
+            String sql = "DELETE FROM staff WHERE staff_id=?";
+            try (
+                    Connection conn = DBConnection.getMySQLConnection();
+                    PreparedStatement ps = conn.prepareStatement(sql);) {
+                ps.setInt(1, aStaff.getStaff_id());
+                ps.executeUpdate();
+                IsDeleted = 1;
+                msg = "Staff has been Deleted";
+                this.clearStaff(aStaff);
+            } catch (Exception e) {
+                LOGGER.log(Level.ERROR, e);
+            }
         }
         FacesContext.getCurrentInstance().addMessage("Save", new FacesMessage(ub.translateWordsInText(BaseName, msg)));
         return IsDeleted;
